Updated repository structure to include additional routers and renamed src to gptuesdays
