+++ gptuesdays/constants.py
