+++ gptuesdays/main.py
@@ -3,17 +3,18 @@ import os
 
 from fastapi import FastAPI, Request, APIRouter
 from dotenv import load_dotenv
-from src import constants
+from gptuesdays import constants
+from gptuesdays.social_media import router as social_media_router 
 
 load_dotenv()
-
-router = APIRouter()
 
 if constants.LOCAL_OR_STAGING_OR_PROD == "local":
     app = FastAPI()
 else:
     app = FastAPI(docs_url=None, redoc_url=None)
 
-@router.get("/api/hello-world/")
+app.include_router(social_media_router.router)
+
+@app.get("/api/hello-world/")
 async def hello_world(request: Request):
     return {"message": "Hello, World! I'm GPTuesday!"}
